pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Probabilistic-Forecasting"
version = "0.0.0"
description = "Probabilistic Forecasting Challenge in winter term 24/25 at KIT"
authors = [{ name = "Moritz Mistol", email = "moritz.mistol@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "darts>=0.31.0",
    "dvc-s3>=3.2.0",
    "dvc>=3.55.2",
    "holidays>=0.59",
    "loguru>=0.7.2",
    "matplotlib>=3.9.2",
    "numpy<2.0.0",
    "omegaconf>=2.3.0",
    "pandas[excel]>=2.2.3",
    "pyarrow>=17.0.0",
    "requests-cache>=1.2.1",
    "seaborn>=0.13.2",
    "sktime>=0.34.0",
    "statsforecast>=1.7.8",
    "tqdm>=4.66.5",
    "u8darts>=0.31.0",
    "xgboost>=2.1.2",
    "shap>=0.46.0",
    "openmeteo-requests>=1.3.0",
    "retry-requests>=2.0.0",
    "lightgbm>=4.5.0",
    "scikit-learn>=1.5.2",
    "catboost>=1.2.7",
]


[tool.uv]
dev-dependencies = [
    "jupyter>=1.1.1",
    "pdoc>=14.7.0",
    "pre-commit>=3.8.0",
    "pytest>=8.3.3",
    "ruff>=0.6.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/probafcst"]


[tool.ruff]
line-length = 88
extend-exclude = ["src/probafcst/__init__.py"]

[tool.ruff.lint.extend-per-file-ignores]
# Also ignore `E402` in all `__init__.py` files.
"__init__.py" = ["E402"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # pydocstyle
    "D",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # pandas related
    "PD",
]
extend-ignore = ["D203", "D213", "E501"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
